Fix implicit conversion between int64_t to int
Summary:
Make conversion explicit, implicit conversion breaks the build
Closes https://github.com/facebook/rocksdb/pull/1589

Differential Revision: D4245158

Pulled By: IslamAbdelRahman

fbshipit-source-id: aaec00d

+++ db/column_family.cc
@@ -571,7 +571,8 @@ int GetL0ThresholdSpeedupCompaction(int level0_file_num_compaction_trigger,
   if (res >= port::kMaxInt32) {
     return port::kMaxInt32;
   } else {
-    return res;
+    // res fits in int
+    return static_cast<int>(res);
   }
 }
 }  // namespace
